reed_py extracted to a separate class
