+++ notebook_v4_to_py.py
@@ -45,7 +45,7 @@
    ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
    WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 """
-from app.Formater import ToPy
+from app.Formatter import ToPy
 from app.NotebookConverter import NotebookConverter
