+++ tests/ReadPyTests.py
@@ -1,0 +1,6 @@
+import unittest
+from app.ReadPy import ReadPy
+
+class ToPyTests(unittest.TestCase):
+    def test_object_creation(self):
+        tp = ReadPy()
